% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clootl_taxonomy_version.R
\name{clootl_taxonomy_version}
\alias{clootl_taxonomy_version}
\title{Version of the clootl taxonomy bundled in the package}
\usage{
clootl_taxonomy_version()
}
\value{
The bundled clootl taxonomy year (an integer).
}
\description{
\code{\link{match_taxonomy}} matches host names against a snapshot of the
\pkg{clootl} (eBird/Clements) avian taxonomy that is bundled with
\code{malaviR}. This returns the taxonomy year of that snapshot.
}
\examples{
clootl_taxonomy_version()
}
\references{
McTavish EJ, Gerbracht JA, Holder MT, Iliff MJ, Lepage D, Rasmussen PC,
Redelings BD, Sanchez Reyes LL, Miller ET (2025). A complete and dynamic tree
of birds. Proceedings of the National Academy of Sciences 122(18):
e2409658122. \doi{10.1073/pnas.2409658122}
}
\seealso{
\code{\link{match_taxonomy}}
}
